검색 프로세스 구조 변경
