--- app/src/main/kotlin/com/example/blogsearcher/app/search/query/BlogSearchApi.kt
@@ -1,5 +1,0 @@
-package com.example.blogsearcher.app.search.query
-
-interface BlogSearchApi {
-    fun search(spec: BlogSearchSpec): BlogSearchResult?
-}
