--- app/src/main/kotlin/com/example/blogsearcher/app/search/query/BlogSearchResult.kt
@@ -1,23 +1,0 @@
-package com.example.blogsearcher.app.search.query
-
-import java.time.LocalDate
-
-class BlogSearchResult(
-    val page: BlogSearchPage,
-    val items: List<BlogSearchItem>
-)
-
-class BlogSearchPage(
-    val total: Int,
-    val pageNo: Int,
-    val sizePerPage: Int,
-    val resultCnt: Int
-)
-
-class BlogSearchItem(
-    val title: String,
-    val content: String,
-    val blogname: String,
-    val postdate: LocalDate,
-    val url: String
-)
