--- app/src/main/kotlin/com/example/blogsearcher/app/search/query/BlogSearchSpec.kt
@@ -1,9 +1,0 @@
-package com.example.blogsearcher.app.search.query
-
-import com.example.blogsearcher.domain.search.vo.Keyword
-import com.example.blogsearcher.domain.search.vo.Page
-
-class BlogSearchSpec(
-    val keyword: Keyword,
-    val page: Page
-)
