+++ domain/src/main/kotlin/com/example/blogsearcher/domain/search/BlogSearchSourceRepository.kt
@@ -1,0 +1,5 @@
+package com.example.blogsearcher.domain.search
+
+interface BlogSearchSourceRepository {
+    fun findAll(): List<BlogSearchSource>
+}
