--- infra/client/src/main/kotlin/com/example/blogsearcher/infra/client/core/factory/RetrofitFactory.kt
@@ -1,8 +1,0 @@
-package com.example.blogsearcher.infra.client.core.factory
-
-interface RetrofitFactory<T> {
-
-    val properties: RetrofitProperties
-
-    fun create(clazz: Class<T>): T
-}
