--- infra/client/src/main/kotlin/com/example/blogsearcher/infra/client/core/factory/RetrofitProperties.kt
@@ -1,9 +1,0 @@
-package com.example.blogsearcher.infra.client.core.factory
-
-data class RetrofitProperties(
-    val baseUrl: String,
-    val connectTimeout: Long = 5L,
-    val readTimeout: Long = 5L,
-    val headers: Map<String, String>,
-    val loggingLevel: String = "BASIC"
-)
