--- infra/client/src/main/kotlin/com/example/blogsearcher/infra/client/kakao/KakaoConstants.kt
@@ -1,10 +1,0 @@
-package com.example.blogsearcher.infra.client.kakao
-
-object KakaoConstants {
-    const val PARAM_QUERY = "query"
-    const val PARAM_SORT = "sort"
-    const val SORT_ACCURACY = "accuracy"
-    const val SORT_RECENCY = "recency"
-    const val PARAM_PAGE = "page"
-    const val PARAM_SIZE = "size"
-}
