--- infra/persistence/build.gradle.kts
@@ -1,15 +1,0 @@
-plugins {
-    id("org.springframework.boot") version "2.7.9"
-    id("io.spring.dependency-management") version "1.0.15.RELEASE"
-    kotlin("plugin.spring") version "1.6.21"
-    kotlin("jvm") version "1.6.21"
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation(project(":domain"))
-    implementation("org.springframework.boot:spring-boot-starter-data-jpa")
-}
