+++ infra/src/main/kotlin/com/example/blogsearcher/infra/retrofit/core/factory/RetrofitFactory.kt
@@ -1,0 +1,8 @@
+package com.example.blogsearcher.infra.retrofit.core.factory
+
+interface RetrofitFactory<T> {
+
+    val properties: RetrofitProperties
+
+    fun create(clazz: Class<T>): T
+}
