+++ infra/src/main/kotlin/com/example/blogsearcher/infra/retrofit/core/factory/RetrofitProperties.kt
@@ -1,0 +1,9 @@
+package com.example.blogsearcher.infra.retrofit.core.factory
+
+data class RetrofitProperties(
+    val baseUrl: String,
+    val connectTimeout: Long = 5L,
+    val readTimeout: Long = 5L,
+    val headers: Map<String, String>,
+    val loggingLevel: String = "BASIC"
+)
