+++ presentation/build.gradle.kts
@@ -18,7 +18,7 @@ tasks.withType<BootJar> {
 dependencies {
     implementation(project(":app"))
     implementation(project(":domain"))
-    implementation(project(":infra:client"))
+    implementation(project(":infra"))
 //    implementation(project(":infra:persistence"))
     implementation("org.springframework.boot:spring-boot-starter-web")
     implementation("org.springframework.boot:spring-boot-starter-validation")
