+++ settings.gradle.kts
@@ -3,4 +3,4 @@ rootProject.name = "blog-searcher"
 include("presentation")
 include("app")
 include("domain")
-include("infra:persistence", "infra:client")
+include("infra")
